+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @prairielearn/ui
 
+## 1.10.0
+
+### Minor Changes
+
+- e2bffd9: Prefer `className` instead of `class`
+
+### Patch Changes
+
+- Updated dependencies [e2bffd9]
+  - @prairielearn/preact-cjs@2.0.0
+
+## 1.9.1
+
+### Patch Changes
+
+- f404bb4: Upgrade all JavaScript dependencies
+
 ## 1.9.0
 
 ### Minor Changes